Added scores redraw, Added level indication

+++ include/figure.h
@@ -1,11 +1,11 @@
 #include <figure_position.h>
 
 struct figure_t {
-    int pos_x;
-    int pos_y;
+    short pos_x;
+    short pos_y;
     ENUM_FIGURES type;
     struct figure_position_t* current_pos;
-    figure_t(ENUM_FIGURES type, figure_position_t* pos, int p_y);
+    figure_t(ENUM_FIGURES type, figure_position_t* pos, short p_y);
     int is_there_collision(figure_position_t* pos);
     int rotate_clockwise();
     int rotate_counterclockwise();
